fix build warnings for buildrom (trivial) Signed-off-by: Stefan Reinauer <stepan@coresystems.de> Acked-by: Stefan Reinauer <stepan@coresystems.de>
git-svn-id: svn://svn.coreboot.org/coreboot/trunk@3439 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ util/buildrom/buildrom.c
@@ -66,15 +66,15 @@ int main(int argc, char *argv[])
 		fatal("stat of infile");
 	if (payloadbuf.st_size > (romsize - size)){
 		fprintf(stderr, "ERROR: payload (%d) + coreboot (%d) - Size is %d bytes larger than ROM size (%d).\n", 
-				payloadbuf.st_size, size, 
-				payloadbuf.st_size+size-romsize,
+				(int)payloadbuf.st_size, size, 
+				(int)(payloadbuf.st_size+size-romsize),
 				romsize);
 		exit(1);
 	}
 
 	printf("Payload: %d coreboot: %d ROM size: %d Left space: %d\n",
-			payloadbuf.st_size, size, romsize,
-			romsize-payloadbuf.st_size-size);
+			(int)payloadbuf.st_size, size, romsize,
+			(int)(romsize-payloadbuf.st_size-size));
 
 	cp = malloc(romsize);
 	if (!cp)
